Move files of the simulator node
#519: Move the files to a package
de.bund.bfr.knime.fsklab.v1_9.simulator to not conflict with the ICPMF
deprecated node.
