+++ de.bund.bfr.knime.fsklab.nodes/src/de/bund/bfr/knime/fsklab/nodes/SimulationSetting.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.apache.commons.lang3.StringUtils;
 import org.knime.core.util.Pair;
-import de.bund.bfr.knime.fsklab.nodes.JSSimulatorViewValue.JSSimulation;
+import de.bund.bfr.knime.fsklab.v1_9.simulator.JSSimulatorViewValue.JSSimulation;
 
 /**
  * Load and save simulation. Used in the simulation configurator node.
